[fix] Don't duplicate current year when it's not the last one
Fixes https://go.epfl.ch/INC0549008 in part.

When e.g. `$LASTDATE` gets moved backwards, but not before one Professor has submitted something, it can happen that the last record by year is not the one from the current year. In this case, the `$yeartodo_done` variable would be (wrongly) reset to False, resulting in a disturbing although harmless behavior of duplicating one year in the search results.